pyproject.toml
[project]
name = "ai-job-application-agent"
version = "0.1.0"
description = "Streamlit-first AI job application workflow with deterministic analysis and supervised agent orchestration."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "streamlit==1.40.1",
    "openai==1.98.0",
    "requests==2.32.4",
    "pypdf==5.9.0",
    "python-docx==1.1.2",
    "pandas==2.2.3",
    "markdown-it-py==3.0.0",
    "playwright==1.48.0",
    "reportlab==3.6.12",
]

[dependency-groups]
dev = [
    "pytest==8.3.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.uv]
package = false
